# Conflicts:
#	docker-compose-ci.yml
#	sources/zafira-services/src/main/java/com/qaprosoft/zafira/services/services/application/emails/TestRunResultsEmail.java

+++ sources/zafira-ws/src/main/java/com/qaprosoft/zafira/ws/security/expressions/UserPermissionEvaluator.java
@@ -59,7 +59,7 @@ public class UserPermissionEvaluator implements IUserPermissionEvaluator
 	{
 		if(authentication != null && targetDomainObject instanceof Long)
 		{
-			return ((JwtUserType)authentication.getPrincipal()).getId() == ((Long) targetDomainObject).longValue();
+			return ((JwtUserType)authentication.getPrincipal()).getId() == (Long) targetDomainObject;
 		}
 		return false;
 	}
